+++ src/components/Modal/index.css
@@ -19,7 +19,6 @@
   position: fixed;
   width: 100%;
   bottom: 0;
-  max-width: 360px;
   padding: 32px 16px;
   border-radius: 8px 8px 0px 0px;
   background: var(--grey-100);
